Code cleanup
git-svn-id: https://ssl.drqueue.org/svn/trunk@1331 94d6b6b3-0ee6-0310-ab95-9b4161196ee2

+++ computer_status.c
@@ -107,9 +107,9 @@ void get_loadavg (uint16_t *loadavg)
 
 	fscanf (f_loadavg,"%f %f %f",&a,&b,&c);
 	
-	loadavg[0] = a * 100;
-	loadavg[1] = b * 100;
-	loadavg[2] = c * 100;
+	loadavg[0] = (uint16_t) a * 100;
+	loadavg[1] = (uint16_t) b * 100;
+	loadavg[2] = (uint16_t) c * 100;
 
 	fclose (f_loadavg);
